Add core-data store types to DataRegistry using module augmentation
